Refactor this package to serve as a conda-provisioning system. (#14)
The idea is to provide a central conda instance for all BioC packages, in the
same vein as biocmake. We inherit some of the environment creation +
configuration utilities from the old conda-based basilisk, to make it easier
for downstream packages to create and cache environments. However, we don't
mandate the use of reticulate here due to GLIBCXX version discrepancies.